alert

+++ resources/views/showUsers.blade.php
@@ -9,7 +9,7 @@
 <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/css/bootstrap.min.css" rel="stylesheet" 
 integrity="sha384-gH2yIJqKdNHPEq0n4Mqa/HGKIhSkIHeL5AyhkYV8i59U5AR6csBvApHHNl/vI1Bx" crossorigin="anonymous">
 <link rel="stylesheet" href="css/showusers.css">
-
+ <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
 </head>
 <body>
     <div class="container">
@@ -28,14 +28,11 @@ integrity="sha384-gH2yIJqKdNHPEq0n4Mqa/HGKIhSkIHeL5AyhkYV8i59U5AR6csBvApHHNl/vI1
                 <td>{{$user->user_name }}</td>
                 <td>{{$user->user_address }}</td>
                 <td>{{$user->user_contact }}</td>
-                <td><a href="/delete/{{ $user->id }}" class="btn btn-danger">Delete User</a></td>
+                <td><a href="/delete/{{ $user->id }}" class="btn btn-danger" onclick= 'return confirm("Delete the user?")'>Delete User</a></td>
                 <td><a href="/edit/{{ $user->id }}" class="btn btn-success">Update User</a></td>
             </tr>
-                
             @endforeach
         </table>
         <a href="{{ url('userpage') }}" class="btn btn-xs btn-info pull-right">Create New User</a>
-
-    </div>
 </body>
 </html>
